+++ internal/provider/resource_compute_virtual_machine.go
@@ -198,8 +198,17 @@ func resourceVirtualMachine() *schema.Resource {
 				Type:     schema.TypeList,
 				Computed: true,
 
-				Elem: &schema.Schema{
-					Type: schema.TypeString,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"id": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"status": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+					},
 				},
 			},
 			"boot_options": {